Subscriber/requirements.txt
Flask
paho-mqtt
xmltodict